update stage and validate workers

+++ worker/workers/archive.py
@@ -20,20 +20,29 @@ def make_tarfile(tarfile_name, source_dir, source_size):
 
 
 def archive(batch):
+    # Tar the batch directory and compute checksum
     scratch_tar_path = make_tarfile(tarfile_name=batch['name'], 
                                     source_dir=batch['paths']['origin'],
                                     source_size=batch['du_size'])
     scratch_tar_path = Path(scratch_tar_path)
+    scratch_digest = utils.checksum(scratch_tar_path)
+
     sda_tar_path = f'{config["paths"]["archive"]}/{batch["name"]}.tar'
-    dataset['paths']['archive'] = sda_tar_path
+    batch['paths']['archive'] = sda_tar_path
     
     print('sda put', str(scratch_tar_path), sda_tar_path)
-
     with utils.track_progress_parallel( progress_fn=hsi_put_progress, 
                                         progress_fn_args=(sda_path, batch['du_size'])):
         sda.put(source=scratch_tar_path, target=sda_tar_path)
     
-    scratch_tar_path.unlink()
+    # validate whether the md5 checksums of local and SDA copies match
+    sda_digest = sda.get_hash(sda_path)
+    if sda_digest == scratch_digest:
+        # file successfully uploaded to SDA, delete the local copy
+        scratch_tar_path.unlink()
+    else:
+        raise Exception(f'Archive failed: Checksums of local {scratch_tar_path} ({scratch_digest}) and SDA {sda_tar_path} ({sda_digest}) do not match')
+    return batch
 
 def tar_progress(tar_path, total_size):
     size = Path(tar_path).stat().st_size
@@ -45,13 +54,13 @@ def hsi_put_progress(sda_path, total_size):
 
 if __name__=='__main__':
     # print(config)
-    # dataset = {
+    # batch = {
     #     'name': 'sentieon_val_7',
     #     'paths': {
     #         'origin': '/N/project/DG_Multiple_Myeloma/share/sentieon_val_7'
     #     }
     # }
-    # dataset = {
+    # batch = {
     #     'name': 'worker',
     #     'paths': {
     #         'origin': '/N/u/dgluser/Carbonate/DGL/worker'
@@ -60,7 +69,7 @@ if __name__=='__main__':
     make_tarfile(tarfile_name='sentieon_val_7_bam', 
                 source_dir='/N/project/DG_Multiple_Myeloma/share/sentieon_val_7/bam', 
                 source_size=371544389559)
-    # archive(dataset, '/N/scratch/dgluser/test/sentieon_val_7.tar')
-    # print(dataset)
+    # archive(batch, '/N/scratch/dgluser/test/sentieon_val_7.tar')
+    # print(batch)
